==> notebooks/workspace/0_extract_data.ipynb <==
#### Notebook purpose
This notebooks aims to collect data from http://eventregistry.org API. We extracted 10k articles related to politics, 10k articles related to sports and 10k articles not related neither to politics nor politics (we'll use this third sample to build the model to classify documents). The selection criterion to filter the 10k documents is the relevance measure returned by Event Registry API.

#### Start session with Event Registry API

from eventregistry import *

key = '' # add your key here

er = EventRegistry(apiKey=key)

#### Specify the data range for the queries

import datetime

end_date = datetime.date(2019, 3, 9)
start_date = end_date - datetime.timedelta(days=30)

#### Specify the informations required for each article

output = ArticleInfoFlags(bodyLen=-1,
                          basicInfo=True,
                          title=True,
                          body=True,
                          location=True,
                          url=False,
                          eventUri=False,
                          authors=False,
                          concepts=False,
                          categories=True,
                          links=False,
                          videos=False,
                          image=False,
                          sentiment=False,
                          dates=False,
                          extractedDates=False,
                          duplicateList=False,
                          originalArticle=False,
                          storyUri=False)

#### Specify the queries to run

q_sports = QueryArticlesIter(dateStart=start_date, 
                             dateEnd=end_date,
                             conceptUri=er.getConceptUri('sports'),
                             categoryUri=er.getCategoryUri('sports'),
                             lang='eng',
                             isDuplicateFilter='skipDuplicates')

q_politics = QueryArticlesIter(dateStart=start_date, 
                               dateEnd=end_date,
                               conceptUri=er.getConceptUri('politics'),
                               categoryUri=er.getCategoryUri('politics'),
                               lang='eng',
                               isDuplicateFilter='skipDuplicates')

q_other = QueryArticlesIter(dateStart=start_date, 
                            dateEnd=end_date,
                            ignoreConceptUri=QueryItems.OR([er.getConceptUri('politics'), er.getConceptUri('sports')]),
                            ignoreCategoryUri=QueryItems.OR([er.getCategoryUri('politics'), er.getCategoryUri('sports')]),
                            lang='eng',
                            isDuplicateFilter='skipDuplicates')

#### Turn the query results into a pandas dataframes and save results

import phq_utils.utils_api as utils

data_path = '../../data/'

%%time 
sports_article_df = utils.build_df_from_query(q_sports, er, 10000, ReturnInfo(articleInfo=output))
sports_article_df.to_csv(f'{data_path}sports_10k.csv', index=False)

%%time
politics_article_df = utils.build_df_from_query(q_politics, er, 10000, ReturnInfo(articleInfo=output))
politics_article_df.to_csv(f'{data_path}politics_10k.csv', index=False)

%%time 
other_article_df = utils.build_df_from_query(q_other, er, 10000, ReturnInfo(articleInfo=output))
other_article_df.to_csv(f'{data_path}other_10k.csv', index=False)
