==> notebooks/workspace/1_topic_modelling.ipynb <==
#### Notebook purpose
This notebooks aims to summarize the sample of sports and politics articles collected on http://eventregistry.org into a finite number of main "topics" using the Latent Dirichlet Allocation (LDA) model. Due to time restrictions we didn't spend time fine-tuning neither the models nor the number of topics but this could have been done using likelihood or perplexity criteria.

#### Load data and set up parameters used in the rest of the notebook

import phq_utils.utils_lda as utils

import pandas as pd

data_path = '../../data/'
modelspath = '../../models/'

sports_df = pd.read_csv(f'{data_path}sports_10k.csv')
politics_df = pd.read_csv(f'{data_path}politics_10k.csv')

from gensim.corpora import Dictionary
from gensim.corpora.mmcorpus import MmCorpus

n_docs = 10000
n_topics = 20
max_freq = 0.4
min_wordcount = int(0.01 * n_docs)

#### Build dictionary and bag of words for sports articles contents

%time docs = [utils.process_text(row['body']) for index, row in sports_df.iterrows()]

dictionary = Dictionary(docs)
dictionary.filter_extremes(no_below=min_wordcount, no_above=max_freq)
dictionary.save(f'{data_path}sports_body.dict')

bows = [dictionary.doc2bow(doc) for doc in docs]
MmCorpus.serialize(f'{data_path}sports_body.bow', bows)

#### Build dictionary and bag of words for politics articles contents

%time docs = [utils.process_text(row['body']) for index, row in politics_df.iterrows()]

dictionary = Dictionary(docs)
dictionary.filter_extremes(no_below=min_wordcount, no_above=max_freq)
dictionary.save(f'{data_path}politics_body.dict')

bows = [dictionary.doc2bow(doc) for doc in docs]
MmCorpus.serialize(f'{data_path}politics_body.bow', bows)

#### Build dictionary and bag of words for sports articles titles

%time docs = [utils.process_text(row['title'] if not pd.isna(row['title']) else '') for index, row in sports_df.iterrows()]

dictionary = Dictionary(docs)
dictionary.filter_extremes(no_below=min_wordcount, no_above=max_freq)
dictionary.save(f'{data_path}sports_title.dict')

bows = [dictionary.doc2bow(doc) for doc in docs]
MmCorpus.serialize(f'{data_path}sports_title.bow', bows)

#### Build dictionary and bag of words for politics articles titles

%time docs = [utils.process_text(row['title'] if not pd.isna(row['title']) else '') for index, row in politics_df.iterrows()]

dictionary = Dictionary(docs)
dictionary.filter_extremes(no_below=min_wordcount, no_above=max_freq)
dictionary.save(f'{data_path}politics_title.dict')

bows = [dictionary.doc2bow(doc) for doc in docs]
MmCorpus.serialize(f'{data_path}politics_title.bow', bows)

#### LDA modeling for politics and sports articles on body and titles

from gensim.models import ldamodel

%%time
for category in ['sports', 'politics']:
    for item in ['body', 'title']:
        dictionary = Dictionary.load(f'{data_path}{category}_{item}.dict')
        corpus = MmCorpus(f'{data_path}{category}_{item}.bow')
        model = ldamodel.LdaModel(corpus=corpus, id2word=dictionary, num_topics=n_topics, random_state=42)
        model.save(f'{models_path}{category}_{item}_{n_docs}_docs_{n_topics}_topics.model')

#### Look at the 10 most important tokens per topic to get an idea of what it represents

import gensim

topic_tokens_top3 = {}
for category in ['sports', 'politics']:
    for item in ['body', 'title']:
        dictionary = Dictionary.load(f'{data_path}{category}_{item}.dict')
        model = gensim.models.LdaModel.load(f'{models_path}{category}_{item}_{n_docs}_docs_{n_topics}_topics.model')
        print(f'Topics for {category} articles {item}:')
        top3_tokens_list = []
        for ix in range(model.num_topics):
            topic_tokens = model.get_topic_terms(ix, 10)
            topic_tokens_list = [dictionary[pair[0]] for pair in topic_tokens]
            print(f'topic {ix} ~ ({", ".join(topic_tokens_list)})')
            top3_tokens_list.append(topic_tokens_list[:3])
        topic_tokens_top3.update({f'{category}_{item}': [' \n'.join(top3) for top3 in top3_tokens_list]})
        print()

#### Project each article in the corresponding topic space and save coordinates

import numpy as np

%%time
for category in ['sports', 'politics']:
    for item in ['body', 'title']:
        corpus = MmCorpus(f'{data_path}{category}_{item}.bow')
        model = gensim.models.LdaModel.load(f'{models_path}{category}_{item}_{n_docs}_docs_{n_topics}_topics.model')
        corpus_rep_in_topic_space = []
        for bow in corpus:
            z = np.zeros(n_topics)
            for topic, weight in model[bow]:
                z[topic] = weight
            corpus_rep_in_topic_space.append(z.tolist())
        df = pd.DataFrame(corpus_rep_in_topic_space)
        df.columns = topic_tokens_top3[f'{category}_{item}']
        df.to_csv(f'{data_path}{category}_{item}_{n_docs}_docs_{n_topics}_topics_lda_weights.csv', index=False)
